``pgtrigger.FSM`` enforces a finite state machine on a field.
The ``pgtrigger.FSM`` trigger allows a user to configure a field and
a set of valid transitions for the field. An error will be raised
if any transitions happen that are not part of the valid transitions
list.

The docs were updated with an example of how to use ``pgtrigger.FSM``.

Type: feature