Optionally change "other" DB name if set at all
Some processes, such as building the documentation, doesn't require a full setup, such as a running database. Hence those builds tend to run without the DATABASE_URL environment variable even set. But `settings.py` assumes it is, as it copies the DB and modifies it's name slightly, which breaks the documentation builds.

Type: trivial