Throw errors on invalid trigger definitions.
Previously triggers were installed with a broad try/except in order to ignore
errors when installing duplicate triggers. This caused invalid triggers to
not be installed with no errors thrown.

The code was updated to catch the specific exception for duplicate triggers
and allow other trigger errors to surface. A failing test case was
added.

Type: bug