Added deferrable triggers
Triggers now have an optional ``timing`` argument. If set, triggers
will be created as "CONSTRAINT" triggers that can be deferred.

When ``timing`` is set to ``pgtrigger.Immediate``, the trigger will
run at the end of a statement. ``pgtrigger.Deferred`` will cause
the trigger to run at the end of the transaction.

Note that deferrable triggers must have both
``pgtrigger.After`` and ``pgtrigger.Row`` values set for the
``when`` and ``level`` attributes.

Type: feature