Add `DatabaseSchemaEditorMixin.create_model()` override to detect and create missing triggers
Django can create tables without migrations if an app is left
'unmigrated' (doesn't have a migrations package in the app) and the
run_syncdb option is provided to the migrate command. This change allows
triggers to be created in this manner.

We override `DatabaseSchemaEditorMixin.create_model()` to detect missing
triggers from one migration state to the next (or reverse if
unapplying), and add them to the database if they're missing.

This also has the side effect of recreating triggers when a migration
that removed some triggers is unapplied, the triggers are now recreated.