Fixes #126 Pass app label, model name into _inject_m2m_depenency_in_proxy
django's implementation of generate_created_proxies uses the project
state to create the operations. Prior to this commit apps.get_model,
since no app_label and model_name were provided, get_model would assume
that `base` was a string. If a model mixin was used, a class would be
passed in instead resulting in a

    AttributeError: type object 'SomeMixin' has no attribute 'split'

By explicitly passing in the app_label and model_name, this line
https://github.com/django/django/blob/main/django/apps/registry.py#L205-L206

is skipped and the migrations are generated as normal.

Type: bug